feat: Refactor, add tests, dry-run mode, and improve error handling
This commit introduces a series of major enhancements to the YouTube Subtitles Manager, transforming it into a more robust, maintainable, and user-friendly application.

The following key improvements have been made:

1.  **Code Refactoring and Modularity:**
    *   The single `yousubv5.py` script has been refactored into a modular structure within a `src` directory.
    *   The code is now organized into logical modules for configuration (`config.py`), YouTube API interactions (`youtube_api.py`), file handling (`file_handler.py`), and the main application logic (`main.py`).
    *   A new entry point, `yousub.py`, has been created at the root of the project.

2.  **"Dry Run" Mode:**
    *   A `--dry-run` flag has been added to all data-modifying commands (`upload`, `update`, `delete`, `process`, `smart-upload`).
    *   When this flag is active, the script simulates the intended actions and prints them to the console without making any actual API calls, providing a crucial safety feature for users.

3.  **Improved Error Handling:**
    *   Generic `except Exception` blocks have been replaced with more specific handlers for `ValueError`, `FileNotFoundError`, `PermissionError`, and `HttpError`.
    *   The script now parses the JSON response from `HttpError` to provide more specific and actionable error messages from the YouTube API.
    *   Authentication error handling has been improved to catch `RefreshError` specifically.

4.  **Automated Testing:**
    *   A `pytest` test suite has been introduced in a new `tests` directory.
    *   Unit tests have been added for core logic, such as language code normalization.
    *   Mocked tests have been created for API functions to verify their behavior (including the `dry_run` mode) without making actual API calls.
