+++ yousub.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from src.main import main
+
+if __name__ == "__main__":
+    main()
